fix: only log if there's a runtime/non-static ipx
